hw: Change hw_get_time_us() to return a 64-bit time instead of 32-bit
A 32-bit counter for microsecond timestamps wraps around within about
an hour. This has the potential problem of causing hangs that are
virtually impossible to reproduce or debug.

+++ src/hw/sdl/hwsdl.c
@@ -54,7 +54,7 @@ void hw_log_error(const char *msg)
 
 /* -------------------------------------------------------------------------- */
 
-uint32_t hw_get_time_us(void)
+int64_t hw_get_time_us(void)
 {
     return SDL_GetTicks() * 1000;
 }
